Replaced the mysql query to delete attachments with wp_delete_attachment()
I replaced the WPDB query with the wp native function wp_delete_attachment(). Mainly it fix a problem because attachments data aren't only into post table and so removing the post will not remove all attachment informations. For example, into wp_postmeta where meta_key = wp_attached_file ) there are informations about the filename.
Also, if there are plugins that hook or filter the wp_delete_attachment() function with the mysql query we do not trigger it (I think we would like to do it instead).